refactor(ast): move fingerprint reference detection
Move fingerprint-variable reference traversal out of task compilation and into the taskfile AST package as Task.ReferencesFingerprintVar. Keep traversal helpers private and add focused coverage for commands, dependencies, preconditions, status, loops, matrices, nested variables, and nil entries.
